+++ src/main/java/org/truffleruby/language/supercall/CallSuperMethodNode.java
@@ -10,19 +10,25 @@
 package org.truffleruby.language.supercall;
 
 import org.truffleruby.core.array.ArrayUtils;
+import org.truffleruby.language.FrameSendingNode;
 import org.truffleruby.language.RubyBaseNode;
 import org.truffleruby.language.arguments.RubyArguments;
 import org.truffleruby.language.dispatch.CallDispatchHeadNode;
 import org.truffleruby.language.methods.CallInternalMethodNode;
 import org.truffleruby.language.methods.InternalMethod;
 
+import com.oracle.truffle.api.Assumption;
 import com.oracle.truffle.api.CompilerDirectives;
+import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
 import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.api.frame.MaterializedFrame;
 import com.oracle.truffle.api.frame.VirtualFrame;
+import com.oracle.truffle.api.nodes.InvalidAssumptionException;
 import com.oracle.truffle.api.object.DynamicObject;
 import com.oracle.truffle.api.profiles.ConditionProfile;
+import com.oracle.truffle.api.utilities.CyclicAssumption;
 
-public abstract class CallSuperMethodNode extends RubyBaseNode {
+public abstract class CallSuperMethodNode extends FrameSendingNode {
 
     private final ConditionProfile missingProfile = ConditionProfile.createBinaryProfile();
 
@@ -50,7 +56,7 @@ public abstract class CallSuperMethodNode extends RubyBaseNode {
         }
 
         final Object[] frameArguments = RubyArguments
-                .pack(null, null, superMethod, null, self, (DynamicObject) block, arguments);
+                .pack(null, getFrameIfRequired(frame), superMethod, null, self, (DynamicObject) block, arguments);
 
         return callMethod(superMethod, frameArguments);
     }